Individualized weapon effects
Every weapon now has some additional effects applied when used to kill,
which are all unique and slightly different.

Weapons have a skill level associated with them which resets to 0 when
equipping a new weapon. Weaponskill is increased by sparring in the
dojo.

It's now possible to miss, and to crit on attacks.

Attack damage is much more varied.

Slimes awarded on kill are based on the total damage the dying player
has taken in their lifetime, making long-lived players more valuable
targets.