Don't allow slime to go negative by attacking
Some weapons have a wef function which increases the slime spent on the
attack, making it more expensive to attack and potentially causing the
attackers slime to go negative if they were close to the shot cost at
the time of the shot.

In this case, slime now goes to zero.